Extract search methods into `Meilisearch::Index::Search` module
